Test branch

+++ lib/list_view_test/models/hotel_detail_model/room_type_model.dart
@@ -17,7 +17,7 @@ class RoomTypeModel {
   final int maxPeopleAllowed;
 
   @JsonKey(name: 'price')
-  final int roomPrice;
+  final double roomPrice;
 
   RoomTypeModel({
     required this.type,
